Updated comments
