+++ package.json
@@ -6,7 +6,7 @@
     "@amsterdam/amsterdam-react-final-form": "^0.1.139",
     "@amsterdam/asc-assets": "^0.36.0",
     "@amsterdam/asc-ui": "^0.36.0",
-    "@amsterdam/wonen-ui": "^1.0.37",
+    "@amsterdam/wonen-ui": "^1.0.38",
     "@material-ui/core": "^4.12.4",
     "@material-ui/icons": "^4.11.3",
     "@reach/router": "^1.3.4",
