+++ src/app/components/cases/TableCases/TableCases.tsx
@@ -2,6 +2,7 @@ import { Table } from "@amsterdam/wonen-ui"
 import navigateTo from "app/routing/navigateTo"
 import getColumns from "./columns"
 import useMediaQuery from "app/hooks/useMediaQuery/useMediaQuery"
+import createResponsiveColumns from "./createPrioritizedColumns"
 
 type Props = {
   data: Components.Schemas.Case[]
@@ -12,29 +13,11 @@ type Props = {
   emptyPlaceholder: string
 }
 
-const columnPriority = [
-  "navigateId",
-  "address.street_name",
-  "address.postal_code",
-  "start_date",
-  "workflows",
-  "reason.name",
-  "last_updated",
-  "id"
-]
-
 const TableCases: React.FC<Props> = ({ data, isBusy, onChange, pagination, sorting, emptyPlaceholder }) => {
   const { windowWidth } = useMediaQuery()
 
   const columns = getColumns(sorting)
-  let prioritizedColumns = columns
-  if (windowWidth < 900) {
-    const prioArray = columnPriority.slice(0, 4)
-    prioritizedColumns = columns.filter(col => prioArray.includes(col.dataIndex))
-  } else if (windowWidth < 1400) {
-    const prioArray = columnPriority.slice(0, 6)
-    prioritizedColumns = columns.filter(col => prioArray.includes(col.dataIndex))
-  }
+  const prioritizedColumns = createResponsiveColumns(columns, windowWidth)
 
   const onClickRow = (data: any) => {
     navigateTo("/zaken/:id", { id: data.id })
